% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiation.R
\name{differentiate}
\alias{differentiate}
\title{Performs statistical test to determine the differentiation level between two sets of gene expression samples.
Performs univariate (rowwise) Welch tests.}
\usage{
differentiate(expressions, control_group, target_group)
}
\arguments{
\item{expressions}{Gene expression matrix: row = gene, column = sample and element = expression level.}

\item{control_group}{Column indexes (samples) that are part of the differentiation control group.}

\item{target_group}{Column indexes (samples) that are part of the differentiation target group.}
}
\value{
List of gene differentiation levels, in the order of the lines of the input matrix.
}
\description{
Performs statistical test to determine the differentiation level between two sets of gene expression samples.
Performs univariate (rowwise) Welch tests.
}
\seealso{
See example of use in the site \url{https://github.com/joseflaviojr/transmod}
}
