% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression_gse48173}
\alias{expression_gse48173}
\title{RNA-Seq data corresponding to experimental study about leukemia.
The table contains 72 human samples: 43 Acute Myeloid Leukemia (AML),
12 Acute Lymphoblastic Leukemia (ALL) and 17 Healthy (HEA).
The samples cover 21865 genes.
Citation: Macrae T, Sargeant T, Lemieux S, Hébert J et al. RNA-Seq reveals
spliceosome and proteasome genes as most consistent
transcripts in human cancer cells. PLoS One, 2013.}
\format{An object of class \code{data.frame} with 21865 rows and 72 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE48173}
}
\usage{
expression_gse48173
}
\description{
RNA-Seq data corresponding to experimental study about leukemia.
The table contains 72 human samples: 43 Acute Myeloid Leukemia (AML),
12 Acute Lymphoblastic Leukemia (ALL) and 17 Healthy (HEA).
The samples cover 21865 genes.
Citation: Macrae T, Sargeant T, Lemieux S, Hébert J et al. RNA-Seq reveals
spliceosome and proteasome genes as most consistent
transcripts in human cancer cells. PLoS One, 2013.
}
\keyword{datasets}
