% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity.R
\name{modularize}
\alias{modularize}
\title{Clustering of adjacent values in a series based on peaks and valleys detection.
Module: series interval with level elevation.}
\usage{
modularize(series, height_min = 0.05)
}
\arguments{
\item{series}{Series of proportional values/levels that have gradual elevations.}

\item{height_min}{Minimum peak height (percentage in relation to the highest value).}
}
\value{
List that numerically indicates in which module each element of the series is.
}
\description{
Clustering of adjacent values in a series based on peaks and valleys detection.
Module: series interval with level elevation.
}
\seealso{
See example of use in the site \url{https://github.com/joseflaviojr/transmod}

Other modularity: \code{\link{select_from_modules}},
  \code{\link{summarize_modules}}
}
