% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity.R
\name{select_from_modules}
\alias{select_from_modules}
\title{Selects elements of a series according to modular levels.
The modules are classified according to the peak level (the higher the better).}
\usage{
select_from_modules(series, modules, select = 100, inspect = 0)
}
\arguments{
\item{series}{Series of proportional values/levels that have gradual elevations.}

\item{modules}{List that numerically indicates in which module each element of the series is.}

\item{select}{Number of elements to select.}

\item{inspect}{Number of modules to consider. 0 = all modules.}
}
\value{
Index (from 1) of each selected element.
}
\description{
Selects elements of a series according to modular levels.
The modules are classified according to the peak level (the higher the better).
}
\seealso{
See example of use in the site \url{https://github.com/joseflaviojr/transmod}

Other modularity: \code{\link{modularize}},
  \code{\link{summarize_modules}}
}
