% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seriation_dias2016}
\alias{seriation_dias2016}
\title{Seriation of human gene network (HI-II-14).
The network was specified in the work "A proteome-scale map of
the human interactome network" (Rolland et al., 2014).
The seriation was specified in the work "A module-based approach
for evaluating differential genome-wide expression profiles" (Dias et al., 2016).}
\format{An object of class \code{character} of length 4303.}
\source{
\url{https://doi.org/10.1109/BRACIS.2016.069}
}
\usage{
seriation_dias2016
}
\description{
Seriation of human gene network (HI-II-14).
The network was specified in the work "A proteome-scale map of
the human interactome network" (Rolland et al., 2014).
The seriation was specified in the work "A module-based approach
for evaluating differential genome-wide expression profiles" (Dias et al., 2016).
}
\keyword{datasets}
