% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity.R
\name{summarize_modules}
\alias{summarize_modules}
\title{Summarizes the main features of each module.}
\usage{
summarize_modules(series, modules)
}
\arguments{
\item{series}{Series of proportional values/levels that have gradual elevations.}

\item{modules}{List that numerically indicates in which module each element of the series is.}
}
\value{
Summary matrix.
}
\description{
Summarizes the main features of each module.
}
\seealso{
See example of use in the site \url{https://github.com/joseflaviojr/transmod}

Other modularity: \code{\link{modularize}},
  \code{\link{select_from_modules}}
}
