% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptogram.R
\name{transcriptogram}
\alias{transcriptogram}
\title{Normalizes gene expression profiles to transcriptogram curves.
The rows of the matrix (genes) must be in the correct order (seriation).}
\usage{
transcriptogram(expressions, window = 251)
}
\arguments{
\item{expressions}{Seriated gene expressions: row = gene, column = sample and element = expression level.}

\item{window}{Window size for calculating the mean of adjacent expression.}
}
\value{
Transcriptograms matrix.
}
\description{
Normalizes gene expression profiles to transcriptogram curves.
The rows of the matrix (genes) must be in the correct order (seriation).
}
\seealso{
See example of use in the site \url{https://github.com/joseflaviojr/transmod}

Other transcriptogram: \code{\link{transcriptogram_series}}
}
