% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmod.R
\docType{package}
\name{transmod}
\alias{transmod}
\alias{transmod-package}
\title{Method to specify list of most relevant genes among
differential expression profiles based on gene network knowledge.
This is done through transcriptogram of gene expression profiles and
analysis of differentially expressed modules.}
\description{
Method to specify list of most relevant genes among
differential expression profiles based on gene network knowledge.
This is done through transcriptogram of gene expression profiles and
analysis of differentially expressed modules.
}
